+++ gitgitgo.go
@@ -2,33 +2,107 @@ package main
 
 import (
 	"os"
+	"os/exec"
 	"phil1436/GitGitGo-CLI/pkg/cmdtool"
 	"phil1436/GitGitGo-CLI/pkg/compiler"
 	"phil1436/GitGitGo-CLI/pkg/logger"
 	"phil1436/GitGitGo-CLI/pkg/subcommands"
 	"phil1436/GitGitGo-CLI/pkg/utils"
+	"runtime"
+	"strings"
 )
 
 func main() {
-
+	// Set default values
+	gitname := GetGitName()
+	utils.PROVIDER = gitname
+	utils.GITHUBNAME = gitname
+	utils.FULLNAME = gitname
 	utils.SetRepoName()
+
+	// Handle global config file
+	homePath := FindCommandDir("gitgitgo")
+	if homePath != "" {
+		utils.HOME = homePath
+		utils.BINARY = FindCommandPath("gitgitgo")
+		compiler.HandleGitGitGoConfigFile(homePath)
+	}
 
 	InitCommands()
 
 	compiler.ExecuteCommand(os.Args, true)
+
+	utils.CheckForUpdate()
 
 	logger.Log("")
 	logger.Log("by Philipp B.")
 	logger.Log("Have a nice day :)")
 }
 
+func GetGitName() string {
+	cmd := exec.Command("git", "config", "--global", "user.name")
+	output, err := cmd.Output()
+	if err != nil {
+		return ""
+	}
+	return strings.TrimSpace(string(output))
+}
+
+func FindCommandDir(command string) string {
+	var cmd *exec.Cmd
+	if runtime.GOOS == "windows" {
+		cmd = exec.Command("where", command)
+	} else {
+		cmd = exec.Command("which", command)
+	}
+
+	output, err := cmd.Output()
+	if err != nil {
+		return ""
+	}
+
+	// On Windows, 'where' may return multiple paths separated by newline
+	// We take the first one in that case
+	path := strings.TrimSpace(strings.Split(string(output), "\n")[0])
+	// get the directory of the command
+	dir := path[:strings.LastIndex(path, string(os.PathSeparator))]
+	return dir
+}
+
+func FindCommandPath(command string) string {
+	var cmd *exec.Cmd
+	if runtime.GOOS == "windows" {
+		cmd = exec.Command("where", command)
+	} else {
+		cmd = exec.Command("which", command)
+	}
+
+	output, err := cmd.Output()
+	if err != nil {
+		return ""
+	}
+
+	// On Windows, 'where' may return multiple paths separated by newline
+	// We take the first one in that case
+	return strings.TrimSpace(strings.Split(string(output), "\n")[0])
+}
+
 func InitCommands() {
-	compiler.AddSubcommand(cmdtool.NewSubcommand([]string{"version", "-v"}, "Get the current version", compiler.PrintVersion))
+	compiler.AddSubcommand(cmdtool.NewSubcommand([]string{"version", "-v"}, "Get the current version", func(attValue []interface{}, fs *cmdtool.FlagSet) bool {
+		s := "GitGitGo-CLI version " + utils.VERSION
+		if utils.PRE_RELEASE {
+			s += " (Pre-Release)"
+		}
+		logger.Log(s)
+		return true
+	}))
 	helpCommad := cmdtool.NewSubcommand([]string{"help", "-h"}, "Help command", compiler.Help)
 	helpCommad.AddAttribute("command", "The command to get help for")
 	compiler.AddSubcommand(helpCommad)
 
 	initCommand := cmdtool.NewSubcommand([]string{"init", "i"}, "Init command", subcommands.Init)
+
+	updateCommand := cmdtool.NewSubcommand([]string{"update"}, "Update the GitGitGo-CLI", utils.Update)
 
 	initCommand.AddAttribute("keywords", "A comma separated list of keywords to search for")
 
@@ -36,7 +110,7 @@ func InitCommands() {
 
 	addCommand.AddAttribute("filename", "The file to add")
 
-	printCommand := cmdtool.NewSubcommand([]string{"print", "p"}, "Print command", subcommands.Print)
+	printCommand := cmdtool.NewSubcommand([]string{"print", "p"}, "Print a file or a parameter", subcommands.Print)
 
 	printCommand.AddAttribute("filename | parametername", "The file or value to print (If no file is given all files will be printed)")
 
@@ -64,7 +138,9 @@ func InitCommands() {
 			utils.REPONAME = value
 			name = "reponame"
 		} else if utils.IsProviderVarName(name) {
-			utils.PROVIDER = value
+			if !utils.SetProvider(value) {
+				return false
+			}
 			name = "provider"
 		} else if utils.IsFullNameVarName(name) {
 			utils.FULLNAME = value
@@ -87,6 +163,11 @@ func InitCommands() {
 		Description: "<path>: Destination directory",
 		Value:       ".",
 	}
+	homeFlag := &cmdtool.Flag{
+		Name:        []string{"home", "h"},
+		Description: "<path>: Specify the home directory to update to",
+		Value:       "",
+	}
 	asFlag := &cmdtool.Flag{
 		Name:        []string{"as"},
 		Description: "<filename>: Give the file a new name",
@@ -95,6 +176,12 @@ func InitCommands() {
 	nameFlagPrint := &cmdtool.Flag{
 		Name:        []string{"name", "n"},
 		Description: ": If enabled only prints the names",
+		Value:       false,
+		BoolFlag:    true,
+	}
+	parameterFlagPrint := &cmdtool.Flag{
+		Name:        []string{"parameter", "pm"},
+		Description: ": If this flag is set all parameters and their values will be printed",
 		Value:       false,
 		BoolFlag:    true,
 	}
@@ -121,7 +208,7 @@ func InitCommands() {
 		Description: "<file1,file2, ... >: Commas separated list of files to ignore",
 		Value:       "",
 	}
-
+	updateCommand.AddFlag(homeFlag)
 	initCommand.AddFlag(destinationFlag)
 	initCommand.AddFlag(forceFlag)
 	initCommand.AddFlag(dryrunFlag)
@@ -131,6 +218,7 @@ func InitCommands() {
 	addCommand.AddFlag(forceFlag)
 	addCommand.AddFlag(dryrunFlag)
 	printCommand.AddFlag(nameFlagPrint)
+	printCommand.AddFlag(parameterFlagPrint)
 	runCommand.AddFlag(forceRunFlag)
 
 	compiler.AddSubcommand(initCommand)
@@ -139,6 +227,7 @@ func InitCommands() {
 	compiler.AddSubcommand(shellCommand)
 	compiler.AddSubcommand(runCommand)
 	compiler.AddSubcommand(setCommand)
+	compiler.AddSubcommand(updateCommand)
 
 	// global flags
 	quietFlag := &cmdtool.Flag{