+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ---
 
+## [v0.1.0](https://github.com/phil1436/GitGitGo-CLI/tree/0.1.0) (2023-12-9)
+
+_PRE-RELEASE_
+
+-   Bug fixes
+-   added `-parameter` flag for the `print` command
+-   `.gitgitgoc` file can now be named `.gggc` as well
+-   seperator support in `.gitgitgoc` file
+-   Parameter set by default to git name
+-   Global config file support
+-   Shell command `mdkir` added
+-   Shell command `reload` added
+-   Command `update` added
+
+---
+
 ## [v0.0.2](https://github.com/phil1436/GitGitGo-CLI/tree/0.0.2) (2023-11-26)
 
 _PRE-RELEASE_